+++ src/main/java/com/yun/service/ItemService.java
@@ -17,7 +17,7 @@ public interface ItemService {
      * @param objectID
      * @return
      */
-    String likeItem(String objectID,Long userID,String likeState);
+    String likeItem(String objectID,Integer userID,String likeState);
 
     /**
      * 根据对象名精确查对象
@@ -25,7 +25,7 @@ public interface ItemService {
      * @param userID
      * @return
      */
-    Item searchItemByName(String objectName,Long userID);
+    Item searchItemByName(String objectName,Integer userID);
 
     /**
      * 根据对象名模糊查对象
@@ -33,7 +33,7 @@ public interface ItemService {
      * @param userID
      * @return
      */
-    List<Item> searchItemsByName(String objectName,Long userID);
+    List<Item> searchItemsByName(String objectName,Integer userID);
 
     /**
      * 根据对象的所属类目查询对象信息
